Safe delete included

+++ src/UseCases/PragaUseCases/excluirPraga.java
@@ -13,10 +13,17 @@ public class excluirPraga {
         System.out.printf("\nDigite o codigo da praga que deseja excluir: ");
         int codigo = scanner.nextInt();
 
-        if(pragaDAO.excluirPraga(codigo)){
-            System.out.println("\n** Praga excluida com sucesso!");
+        System.out.print("Você tem certeza que deseja excluir? (S/N): ");
+        scanner.nextLine();
+        String confirmacao = scanner.nextLine();
+        if(confirmacao.equals("S")){
+            if(pragaDAO.excluirPraga(codigo)){
+                System.out.println("\n** Praga excluida com sucesso!");
+            }else{
+                System.out.println("\n** Praga não encontrada. Liste todas para obter seus respectivos codigos.");
+            }
         }else{
-            System.out.println("\n** Praga não encontrada. Liste todas para obter seus respectivos codigos.");
+            System.out.println("\n** Ação cancelada!");
         }
     }
 }
